requirements.txt
pytest==9.0.2
praw
requests
psycopg2-binary
pgvector
apache-airflow==2.10.4
streamlit
langgraph
sentence-transformers
plotly
python-dotenv
kafka-python==2.0.2
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
pydantic==2.7.1
dbt-core==1.8.1
dbt-duckdb==1.8.4
redis==5.0.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
openai>=1.30.0
